requirements.txt
asttokens==3.0.0
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.23.0
azure-keyvault-secrets==4.10.0
azure-mgmt-compute==34.1.0
azure-mgmt-core==1.5.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==24.0.0
azure-mgmt-storage==23.0.0
azure-storage-blob==12.25.1
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
distlib==0.3.9
executing==2.1.0
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.2
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
nest-asyncio==1.6.0
numpy==2.2.1
oauthlib==3.3.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
pyzmq==26.2.0
requests==2.32.4
requests-oauthlib==2.0.0
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2024.2
urllib3==2.5.0
virtualenv==20.31.2
wcwidth==0.2.13
Werkzeug==3.1.3
Flask
azure-identity
azure-keyvault-secrets
azure-storage-blob
psycopg2-binary
